requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
git+https://github.com/Ricky294/trader.git@develop
git+https://github.com/Ricky294/cryptodata.git@develop
numpy
pandas
python-binance
torch==1.10.1+cu113
plotly
dash
tqdm
scikit-learn
pyyaml
